refactor: split response_style under the 20 KB gate; grandfather engine.py with a plan
test_v06_module_size_gate was red on two real CLAUDE.md rule #5
violations, neither introduced by recent work:

  core/response_style.py   22,036 bytes  (984 over → 1,556 over)
  core/reasoning/engine.py 21,464 bytes

**response_style.py is split**, not grandfathered — the file's own
GRANDFATHERED comment records an anti-creep rule, and splitting is the
convention. The StylePreset dataclass and the three preset bodies (12 KB
of the 22) move to core/response_style_presets.py; response_style.py
keeps the resolver and re-exports every public name, so the eleven call
sites across core/reasoning/*, answer_style_classifier, memory/extractor
and routes/query are untouched. Import direction is one-way, so no cycle.

    response_style.py           4,383 bytes
    response_style_presets.py  19,065 bytes

Verified equivalent rather than assumed: the pre-split module is loaded
out of git alongside the new one and compared field-for-field. All three
presets fingerprint identically (NATURAL 3,632 chars, TERSE 967,
DETAILED 1,305), the five module constants match, and resolve_style
agrees on eight inputs including the env-var path, an unknown id and
whitespace/case variants. Re-exports are the same objects, not copies.
80 passed across the six suites that touch response_style.

**engine.py is grandfathered with a split plan**, because splitting it
here would be the wrong kind of change. It is core/reasoning, so rule #2
requires STEP 7 bench numbers and a Quality Delta Card — and bench.py
needs a live server plus Ollama, neither of which exists in a session
container. Shipping an unmeasured refactor of the hottest path to make a
gate green is exactly what that rule exists to prevent. The entry names
the seam an operator who can run the bench should use (the mode dispatch
block, the way pipeline_synth was lifted) and notes that the structural
test helpers already absorb either shape.

One bug caught during the split: `@dataclass(frozen=True)` sat one line
above the class and was left behind by the first cut, so StylePreset
came out a plain class and every preset construction raised TypeError.
Caught by importing, not by review.

Quality delta: exempt (label: fix) — response_style is not
core/retrieval, core/graph or core/reasoning, and no reasoning line
changed.
